requirements.txt
addict==2.2.1
azure-devops==6.0.0b2
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isodate==0.6.0
msrest==0.6.13
oauthlib==3.1.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
